+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.2.2 ()
+
+### Added
+
+### Changed
+- move function `format_textheader_string` to utils.py [issue #28](https://github.com/anthonytorlucci/segytools/issues/28)
+- move function `read_trace_data` to utils.py [issue #28](https://github.com/anthonytorlucci/segytools/issues/28)
+
 ## v0.2.1 (2023-07-24)
 
 ### Added
